# src/backend/user-service/openapi.yaml
openapi: 3.0.0
info:
  title: User Management API
  description: REST API for user authentication, management, and 2FA with role-based access control.
  version: 1.1.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000/api
    description: Development server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: authToken
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        avatar:
          type: string
          nullable: true
        twoFactor:
          type: object
          nullable: true
          properties:
            method:
              type: string
              enum: [email, totp]
        role:
          type: string
          enum: [user, admin]
      example:
        id: 1
        username: johndoe
        email: john@example.com
        twoFactor: { method: "email" }
        role: user
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string, minLength: 4 }
        email: { type: string, format: email }
        password: { type: string, minLength: 4 }
    Verify2FARequest:
      type: object
      required: [code]
      properties:
        code: 
          type: string
          description: 6-digit verification code
...
    Friend:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        avatar: { type: string }
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful (or 2FA required)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  requires2FA: { type: boolean }
                  user: 
                    $ref: '#/components/schemas/User'
                  devCode: 
                    type: string
                    description: Development only code (if 2FA required)
        '401':
          description: Invalid credentials
  /auth/login/verify:
    post:
      tags: [Authentication]
      summary: Verify 2FA code during login
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Verify2FARequest'
      responses:
        '200':
          description: 2FA Verified, full session granted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  user:
                    $ref: '#/components/schemas/User'
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
        '400':
          description: Validation error
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh Access Token
      description: Uses the httpOnly refreshToken cookie to issue a new authToken.
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid or missing refresh token
...
  /users/friends:
    get:
      tags: [Friends]
      summary: Get all friends
      security:
        - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  friends:
                    type: array
                    items:
                      $ref: '#/components/schemas/Friend'
...